// src/types/clinic.d.ts
	interface IMedicalRecord {
		_id: string;
		type: string;
		uploadedBy: string;
		url: string;
		comment: string | null;
		vaccinationDate: string | null;
		nextVaccinationDate: string | null;
		vaccinatedOnDate: string | null;
		followUpDate: string | null;
		followUpType: string | null;
		vaccineName: string | null;
		notificationCount: number;
		active: boolean;
		parent: ICommonTypes.IParent;
		clinic: ICommonTypes.IClinic;
		pet: ICommonTypes.IPet;
		createdAtIST: string;
		updatedAtIST: string;
		recordId: string;
		createdAtUTC: string;
		updatedAtUTC: string;
	}
// ...
	interface IParent {
		name: string;
		mobile: string;
		parentId: string;
	}